Fix edge case where unblocking workflow doesn't enqueue job event (#132)
* Fix edge case where unblocking workflow doesn't enqueue job event

There is an edge case when we unblock all failed instances in batches and default batch size = 100, the only failed instance was unblocked in first while loop but didn't enqueue a job event because the condition requires unblocked count > 0 (in 1st loop this value is not updated and is still 0). This PR updated the if condition to account for this case.

* improve comment